refactor(adapter-openclaw): extract agent + workspace config into agent-config.ts
Eighth cut of the runtime.ts god-file split. Moves the 19 agent/workspace
config helpers into a new agent-config.ts: the openclaw.json mutators
(upsertOpenClawAgentConfig, updateOpenClawAgentIdentity, updateAgentAllowlist,
removeOpenClawAgentConfig + the artifact teardown removeOpenClawAgent*/
removeOpenClawOwnedPath), the writeOpenClawConfig writer, agentModelPrimary/
openClawAgentsList/agentToRuntime, the workspace+skill file readers
(getWorkspacePath, isSafeWorkspaceFile, isSafeSkillFilePath, readSkillTree,
readWorkspaceRootFile), readGatewayToken, and the identity parsers
(resolveRole, matchIdentityField).

Reads/writes go through the config + home + main-agent + cron-store siblings,
so no cycle back to runtime. The class's agents/skills methods import back the
13 helpers they call; the internal-only ones stay module-private. runtime.ts
shed its now-unused materializeImplicitMainAgent/OpenClawAgent/OpenClawConfig/
getOpenClawHome/tryGetMainAgentId imports.

Pure relocation — bodies byte-for-byte unchanged. runtime.ts 1,992 -> 1,744.
Verified: typecheck, eslint (changed files), adapter suite 115-0 (--isolate),
full suite 5124-0, binary build (stamps reverted), boot smoke (adapter loads,
schedule + team plugins activate, both list routes -> 200; team exercises the
moved agent helpers via the runtime adapter).
