+++ packages/adapter-openclaw/src/runtime.ts
@@ -37,14 +37,10 @@ import {
   agentListFrom,
   findAgentById,
   getAgentList,
-  materializeImplicitMainAgent,
   readOpenClawConfig,
   resetOpenClawConfigCache,
-  type OpenClawAgent,
-  type OpenClawConfig,
 } from './config'
-import { getOpenClawHome, getOpenClawPath } from './home'
-import { tryGetMainAgentId } from './main-agent'
+import { getOpenClawPath } from './home'
 import type { OpenClawRuntimeAdapterOptions } from './index'
 import { OpenClawApprovalGatewayClient } from './approval-gateway'
 import { OpenClawGatewayRpcClient } from './gateway-rpc'
@@ -80,6 +76,13 @@ import {
   OPENCLAW_SESSION_ACTIVITY_POLL_MS,
   watchOpenClawSessionActivity, createOpenClawSessionActivityCursor, openClawCliSessionId,
 } from './session-activity'
+import {
+  writeOpenClawConfig, upsertOpenClawAgentConfig,
+  updateOpenClawAgentIdentity, updateAgentAllowlist, removeOpenClawAgentConfig,
+  removeOpenClawAgentArtifacts, removeOpenClawAgentCronArtifacts,
+  agentToRuntime, getWorkspacePath, readGatewayToken, isSafeWorkspaceFile, isSafeSkillFilePath,
+  readSkillTree,
+} from './agent-config'
 // Re-exported so the session-store-cache test's `from '.../runtime'` path stays stable.
 export {
   SESSION_STORE_CACHE_MAX, __readSessionStoreCachedForTest,
@@ -1535,141 +1538,6 @@ function openClawChildEnv(): NodeJS.ProcessEnv {
   }
 }
 
-function writeOpenClawConfig(config: Record<string, unknown>): void {
-  mkdirSync(getOpenClawHome(), { recursive: true })
-  writeFileSync(getOpenClawPath('openclaw.json'), JSON.stringify(config, null, 2), 'utf-8')
-  resetOpenClawConfigCache()
-}
-
-function agentModelPrimary(model: OpenClawAgent['model']): string | undefined {
-  if (typeof model === 'string') return model
-  return model?.primary
-}
-
-function openClawAgentsList(config: OpenClawConfig): OpenClawAgent[] {
-  config.agents ??= {}
-  config.agents.list ??= []
-  return config.agents.list
-}
-
-function upsertOpenClawAgentConfig(input: {
-  id: string
-  name: string
-  workspace: string
-  model?: string
-  emoji?: string
-}): void {
-  const config: OpenClawConfig = readOpenClawConfig() ?? {}
-  const list = openClawAgentsList(config)
-  const existing = list.find((agent) => agent.id === input.id)
-  const agentDir = getOpenClawPath('agents', input.id, 'agent')
-  const identity = input.name || input.emoji
-    ? {
-        ...(existing?.identity ?? {}),
-        ...(input.name ? { name: input.name } : {}),
-        ...(input.emoji ? { emoji: input.emoji } : {}),
-      }
-    : existing?.identity
-
-  if (!existing && list.length === 0 && input.id !== 'main') {
-    list.push({ id: 'main' })
-  }
-
-  const next = {
-    ...(existing ?? { id: input.id }),
-    name: input.name,
-    workspace: input.workspace,
-    agentDir,
-    ...(input.model ? { model: input.model } : {}),
-    ...(identity ? { identity } : {}),
-  }
-
-  if (existing) Object.assign(existing, next)
-  else list.push(next)
-
-  mkdirSync(agentDir, { recursive: true })
-  mkdirSync(join(agentDir, 'sessions'), { recursive: true })
-  mkdirSync(input.workspace, { recursive: true })
-  writeOpenClawConfig(config as unknown as Record<string, unknown>)
-}
-
-function updateOpenClawAgentIdentity(agentId: string, input: { name?: string; emoji?: string }): void {
-  const config = readOpenClawConfig()
-  const agent = config?.agents?.list?.find((entry) => entry.id === agentId)
-  if (!agent) throw new Error(`Agent not found: ${agentId}`)
-  agent.identity = {
-    ...(agent.identity ?? {}),
-    ...(input.name ? { name: input.name } : {}),
-    ...(input.emoji ? { emoji: input.emoji } : {}),
-  }
-  writeOpenClawConfig(config as unknown as Record<string, unknown>)
-}
-
-function updateAgentAllowlist(agentId: string, updater: (current: string[]) => string[]): void {
-  const config = readOpenClawConfig()
-  const agent = agentId === 'main' && config
-    ? materializeImplicitMainAgent(config)
-    : config?.agents?.list?.find((entry) => entry.id === agentId)
-  if (!agent) throw new Error(`Agent not found: ${agentId}`)
-  agent.subagents ??= {}
-  agent.subagents.allowAgents = updater(agent.subagents.allowAgents ?? [])
-  writeOpenClawConfig(config as unknown as Record<string, unknown>)
-}
-
-function removeOpenClawAgentConfig(agentId: string): void {
-  const config = readOpenClawConfig()
-  const agents = config?.agents?.list
-  if (!config?.agents || !agents) return
-
-  let changed = false
-  const filtered = agents.filter((agent) => agent.id !== agentId)
-  if (filtered.length !== agents.length) {
-    config.agents.list = filtered
-    changed = true
-  }
-
-  for (const agent of config.agents.list ?? []) {
-    const allowAgents = agent.subagents?.allowAgents
-    if (!allowAgents?.includes(agentId)) continue
-    agent.subagents!.allowAgents = allowAgents.filter((id) => id !== agentId)
-    changed = true
-  }
-
-  if (changed) writeOpenClawConfig(config as unknown as Record<string, unknown>)
-}
-
-function removeOpenClawAgentArtifacts(agentId: string, workspace: string): void {
-  removeOpenClawOwnedPath(workspace)
-  removeOpenClawOwnedPath(getOpenClawPath('agents', agentId))
-}
-
-function removeOpenClawAgentCronArtifacts(agentId: string): void {
-  const store = readCronStore()
-  const jobs = store.jobs ?? []
-  const removedJobIds = new Set<string>()
-  const keptJobs = jobs.filter((job) => {
-    const matches = job.agentId === agentId
-      || job.sessionTarget === agentId
-      || job.sessionTarget === `agent:${agentId}`
-    if (matches && job.id) removedJobIds.add(job.id)
-    return !matches
-  })
-  if (keptJobs.length === jobs.length) return
-
-  writeCronStore({ ...store, jobs: keptJobs })
-  for (const jobId of removedJobIds) {
-    removeOpenClawOwnedPath(getOpenClawPath('cron', 'runs', `${jobId}.jsonl`))
-  }
-}
-
-function removeOpenClawOwnedPath(path: string | undefined): void {
-  if (!path) return
-  const home = resolve(getOpenClawHome())
-  const target = resolve(path)
-  if (target === home || !target.startsWith(`${home}${sep}`)) return
-  rmSync(target, { recursive: true, force: true })
-}
-
 function splitChannelRef(channelId: string, metadata: RuntimeMetadata | undefined): { channel: string; target?: string } {
   const explicitTarget = metadataValue(metadata, 'target') ?? metadataValue(metadata, 'channelTarget')
   if (explicitTarget) return { channel: channelId, target: explicitTarget }
@@ -1873,120 +1741,4 @@ export async function* mergeChatStreams(
   } finally {
     stopSecondary()
   }
-}
-
-function agentToRuntime(agent: NonNullable<ReturnType<typeof findAgentById>>): RuntimeAgent {
-  return {
-    id: agent.id,
-    name: agent.identity?.name ?? agent.name ?? agent.id,
-    role: resolveRole(agent.id),
-    model: agentModelPrimary(agent.model),
-    status: 'active',
-    metadata: {
-      emoji: agent.identity?.emoji ?? '',
-      workspacePath: getWorkspacePath(agent.id),
-      subagentAllowAgents: agent.subagents?.allowAgents ?? null,
-    },
-  }
-}
-
-function getWorkspacePath(agentId: string): string {
-  const config = readOpenClawConfig()
-  const agent = config?.agents?.list?.find((entry) => entry.id === agentId)
-  if (agent?.workspace) return agent.workspace
-  if (agentId === tryGetMainAgentId()) {
-    return config?.agents?.defaults?.workspace ?? join(getOpenClawHome(), 'workspace')
-  }
-  return join(getOpenClawHome(), 'workspaces', agentId)
-}
-
-function readGatewayToken(): string | null {
-  const config = readOpenClawConfig() as { gateway?: { auth?: { token?: unknown } } } | null
-  const token = config?.gateway?.auth?.token
-  return typeof token === 'string' && token.length > 0 ? token : null
-}
-
-function isSafeWorkspaceFile(path: string): boolean {
-  return !path.includes('..') && !path.startsWith('/') && !path.includes('\\')
-}
-
-function isSafeSkillFilePath(path: string): boolean {
-  return Boolean(path)
-    && !path.startsWith('/')
-    && !path.includes('\\')
-    && !path.split('/').some((part) => part === '..' || part === '')
-}
-
-function readSkillTree(root: string): Record<string, string> {
-  const files: Record<string, string> = {}
-  const walk = (dir: string, prefix = ''): void => {
-    for (const entry of readdirSync(dir, { withFileTypes: true })) {
-      if (entry.isSymbolicLink()) continue
-      const rel = prefix ? `${prefix}/${entry.name}` : entry.name
-      const abs = join(dir, entry.name)
-      if (entry.isDirectory()) {
-        walk(abs, rel)
-      } else if (entry.isFile()) {
-        if (entry.name === '.installedBy' || entry.name === '.userEdited') continue
-        files[rel] = readFileSync(abs, 'utf-8')
-      }
-    }
-  }
-  try {
-    walk(root)
-  } catch {
-    return {}
-  }
-  return files
-}
-
-function readWorkspaceRootFile(agentId: string, filename: string): string | null {
-  if (!isSafeWorkspaceFile(filename)) return null
-  try {
-    return readFileSync(join(getWorkspacePath(agentId), filename), 'utf-8')
-  } catch {
-    return null
-  }
-}
-
-function matchIdentityField(identity: string, key: string): string | null {
-  const inlineRe = new RegExp(
-    `^\\s*[-*]?\\s*\\*{0,2}${key}\\*{0,2}\\s*:\\s*\\*{0,2}\\s*(.+?)\\s*\\*{0,2}\\s*$`,
-    'mi',
-  )
-  const inline = identity.match(inlineRe)
-  if (inline) {
-    const value = inline[1].trim().replace(/^\*+|\*+$/g, '').trim()
-    if (value.length > 0) return value
-  }
-  const heading = identity.match(new RegExp(`^#{1,6}\\s+${key}\\s*$\\n+([^\\n]+)`, 'mi'))
-  if (heading) {
-    const value = heading[1].trim().replace(/^\*+|\*+$/g, '').trim()
-    if (value.length > 0) return value
-  }
-  return null
-}
-
-function resolveRole(agentId: string): string {
-  const identity = readWorkspaceRootFile(agentId, 'IDENTITY.md')
-  if (identity) {
-    const role = matchIdentityField(identity, 'Role')
-    if (role) return role
-    const vibe = matchIdentityField(identity, 'Vibe')
-    if (vibe) return vibe
-  }
-
-  const soul = readWorkspaceRootFile(agentId, 'SOUL.md')
-  if (soul) {
-    const firstLine = soul.split('\n').find((line) => line.startsWith('You are ') || line.startsWith('# '))
-    if (firstLine) {
-      const dashPart = firstLine.split('—')[1] || firstLine.split('-')[1]
-      if (dashPart) {
-        const role = dashPart.replace(/\.\s*$/, '').trim()
-        if (role.length > 0 && role.length < 60) return role
-      }
-    }
-  }
-
-  return agentId === tryGetMainAgentId() ? 'Orchestrator' : ''
 }
